--- src/services/ContributingDeveloperAuditService/index.ts
@@ -1,2 +1,0 @@
-export * from "./scm";
-export * from "./ContributingDeveloperAuditService";
