+++ src/services/ContributorAuditService/index.ts
@@ -1,0 +1,2 @@
+export * from "./scm";
+export * from "./ContributorAuditService";
